Add files via upload
ImageJ macro, to create a stitching of several images, over a large amount of time point. This macro is especially useful to create a video of a timelapse made out of several positions.
Name your files in the fashion "position_timepoint" and import this image sequence within imageJ. Start the macro and indicate the number of columns and lines you wish to have. The macro will automatically calculate the number of slides necessary in the new image sequence and paste all image after each other in the new image sequence.